% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comex_stat_tabelas.R
\name{show_available_tables}
\alias{show_available_tables}
\title{Mostrar Tabelas de Referência Disponíveis}
\usage{
show_available_tables()
}
\value{
Imprime no console as informações das tabelas disponíveis. Não retorna nenhum objeto.
}
\description{
Exibe uma lista das tabelas de referência de Comércio Exterior que podem ser
carregadas usando a função \code{load_reference_tables()}, incluindo seus códigos
e descrições.
}
\examples{
\dontrun{
show_available_tables()
}
}
